рефакторинг
пересобрал get и getAll в отдельные функции
