+++ src/prod/index.ts
@@ -1,21 +1,7 @@
-import docsGet from './docsGet';
-import docsGetAll from './docsGetAll';
-import examsGet from './examsGet';
-import examsGetAll from './examsGetAll';
-import teachersGet from './teachersGet';
-import teachersGetAll from './teachersGetAll';
-import testsGet from './testsGet';
-import testsGetAll from './testsGetAll';
+// POST
 import docsPost from './docsPost';
+// GET
+import collectionGet from './collectionGet';
+import collectionGetAll from './collectionGetAll';
 
-export {
-  docsGet,
-  docsGetAll,
-  examsGet,
-  examsGetAll,
-  teachersGet,
-  teachersGetAll,
-  testsGet,
-  testsGetAll,
-  docsPost,
-};
+export { docsPost, collectionGet, collectionGetAll };
